requirements.txt
goose3==3.1.12
matplotlib==3.6.3
numpy==1.24.2
pandas==1.5.3
plotly==5.13.0
python-dotenv==1.0.0
Requests==2.31.0
streamlit==1.21.0
vaderSentiment==3.3.2
vaderSentiment==3.3.2